experimenting

+++ examples/llm/tech_qa.py
@@ -255,7 +255,7 @@ def train(args, data_lists):
     gnn = GAT(in_channels=768, hidden_channels=hidden_channels,
               out_channels=1024, num_layers=num_gnn_layers, heads=4)
     if args. llm_generator_mode == "full":
-        llm = LLM(model_name=args.llm_generator_name, dtype=torch.float32)
+        llm = LLM(model_name=args.llm_generator_name)
         model = GRetriever(llm=llm, gnn=gnn)
     elif args. llm_generator_mode == "lora":
         llm = LLM(model_name=args.llm_generator_name, dtype=torch.float32)
